Fix 32bits warnings

+++ source/system/include/Platform.h
@@ -45,4 +45,4 @@
 /**User litteral for size_t (ie: size_t s = 10_z)
  * @return The integer value as a size_t
  */
-constexpr size_t operator"" _z(unsigned long long int n) { return n; }
+constexpr size_t operator"" _z(unsigned long long int n) { return size_t(n); }
